Created createCard script to separate concerns, properly called functions with event listeners on buttons
