+++ index.html
@@ -16,7 +16,7 @@
             </div>
         <div id="playercard" class="playercard"></div>
         <div class="buttons">
-            <button id="shuffle" onclick="shuffle()">Shuffle & Deal</button>
+            <button id="shuffle">Shuffle & Deal</button>
             <br>
             <button id="playButton" onclick="play()" disabled>Play</button>
             <br>
@@ -38,6 +38,6 @@
         <h2 id="win">You win!</h2>
         <h2 id="lose">Computer wins!</h2>
     </main>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
 </html>
