+++ script.js
@@ -1,3 +1,6 @@
+// Import createCard() from createCard.js
+import createCard from './createCard';
+
 // Initializes the entire deck, your deck, and the computer's deck as empty arrays
 const fullDeck = [];
 const playerDeck = [];
@@ -7,9 +10,16 @@ const computerDeck = [];
 const playerPlay = [];
 const computerPlay = [];
 
+// Adds event listeners to buttons here instead of onclick() functions in index.html
+const shuffleButton = document.getElementById('shuffle');
+shuffleButton.addEventListener('click', shuffle());
+const playButton = document.getElementById('play');
+playButton.addEventListener('click', play());
+
 // Establishes suits and values of cards in a standard 52-card deck
 const suits = ['clubs', 'diamonds', 'hearts', 'spades'];
 const values = ['2', '3', '4', '5', '6', '7', '8', '9', '10', 'J', 'Q', 'K', 'A'];
+
 
 // Creates and shuffles the deck using the Fisher-Yates Algorithm
 function shuffle() {
@@ -62,41 +72,42 @@ function play() {
     console.log(playerDeck.length, computerDeck.length);
     // Prevents you from clicking the Play button before the play is resolved
     document.getElementById('playButton').disabled = true;
-    // Creates a container div with the fullcard class under playerCardArea
-    const playerCardArea = document.getElementById('playercard');
-    const cardContainer = document.createElement('div');
-    cardContainer.setAttribute('class', 'fullcard');
-    playerCardArea.appendChild(cardContainer);
-    // Creates the card image with the cardImg class under cardContainer
-    var cardImg = document.createElement('img');
-    cardImg.setAttribute('class', 'cardImg');
-    cardImg.src = 'images/blank_card.png';
-    cardContainer.appendChild(cardImg);
-    // Creates the card text with the cardText class under cardContainer
-    var cardText = document.createElement('h3');
-    cardText.setAttribute("class", "cardText");
-    cardContainer.appendChild(cardText);
-    // Sets the card value, suit, and text color
-    for (i = 0; i < playerPlay.length; i++) {
-        var card = playerPlay[i];
-        cardContainer.id = `${card.Value} of ${card.Suit}`;
-        switch (true) {
-            case (card.Suit === "clubs"):
-                cardText.textContent = `${card.Value}\u2663`;
-                cardText.style.color = "black";
-                break;
-            case (card.Suit === "diamonds"):
-                cardText.textContent = `${card.Value}\u2666`;
-                cardText.style.color = "red";
-                break;
-            case (card.Suit === "hearts"):
-                cardText.textContent = `${card.Value}\u2665`;
-                cardText.style.color = "red";
-                break;
-            case (card.Suit === "spades"):
-                cardText.textContent = `${card.Value}\u2660`;
-                cardText.style.color = "black";
-                break;
-        }
-    }
+    createCard();
+    // // Creates a container div with the fullcard class under playerCardArea
+    // const playerCardArea = document.getElementById('playercard');
+    // const cardContainer = document.createElement('div');
+    // cardContainer.setAttribute('class', 'fullcard');
+    // playerCardArea.appendChild(cardContainer);
+    // // Creates the card image with the cardImg class under cardContainer
+    // var cardImg = document.createElement('img');
+    // cardImg.setAttribute('class', 'cardImg');
+    // cardImg.src = 'images/blank_card.png';
+    // cardContainer.appendChild(cardImg);
+    // // Creates the card text with the cardText class under cardContainer
+    // var cardText = document.createElement('h3');
+    // cardText.setAttribute("class", "cardText");
+    // cardContainer.appendChild(cardText);
+    // // Sets the card value, suit, and text color
+    // for (i = 0; i < playerPlay.length; i++) {
+    //     var card = playerPlay[i];
+    //     cardContainer.id = `${card.Value} of ${card.Suit}`;
+    //     switch (true) {
+    //         case (card.Suit === "clubs"):
+    //             cardText.textContent = `${card.Value}\u2663`;
+    //             cardText.style.color = "black";
+    //             break;
+    //         case (card.Suit === "diamonds"):
+    //             cardText.textContent = `${card.Value}\u2666`;
+    //             cardText.style.color = "red";
+    //             break;
+    //         case (card.Suit === "hearts"):
+    //             cardText.textContent = `${card.Value}\u2665`;
+    //             cardText.style.color = "red";
+    //             break;
+    //         case (card.Suit === "spades"):
+    //             cardText.textContent = `${card.Value}\u2660`;
+    //             cardText.style.color = "black";
+    //             break;
+    //     }
+    // }
 }
